requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.2.0
auditlog3==1.0.1
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
cssselect2==0.3.0
daphne==2.5.0
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-annoying==0.10.6
django-bootstrap-modal-forms==2.0.0
django-braces==1.14.0
django-colorfield==0.3.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-rated-reviews==0.6.0
django-star-ratings==0.9.1
django-webtest==1.7.7
django-widget-tweaks==1.4.8
django-yandex-money==1.1.3
h11==0.9.0
h2==4.0.0
hpack==4.0.0
html5lib==1.1
httptools==0.1.1
Hypercorn==0.11.1
hyperframe==6.0.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kombu==4.6.11
lxml==4.5.2
oauthlib==3.1.0
pdfkit==0.6.1
phonenumbers==8.12.7
Pillow==7.2.0
pkg-resources==0.0.0
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
South==0.8.4
sqlparse==0.3.1
swapper==1.1.2.post1
tinycss2==1.0.2
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
users==1.0.dev0
uvicorn==0.11.8
uvloop==0.14.0
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
websockets==8.1
wsproto==0.15.0
zope.interface==5.1.2
